feat: 챗봇 레이아웃 컴포넌트 분리 외 #15
라우트 별 렌더링 여부 조정
