+++ src/App.tsx
@@ -1,10 +1,7 @@
 import { routers } from "@src/Router";
-import { useState } from "react";
 import { RouterProvider } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 
-import Chat from "@src/components/Chat/Chat";
-import ChatButton from "@src/components/Chat/ChatButton";
 import ThemeToggle from "@src/components/Toggle/ThemeToggle";
 
 import useModeTheme from "@src/hooks/useModeTheme";
@@ -15,11 +12,6 @@ import { DARK_MODE_THEME, LIGHT_MODE_THEME } from "@src/@styles/theme";
 
 function App() {
   const { theme } = useModeTheme();
-  const [isChatVisible, setIsChatVisible] = useState(false);
-
-  const toggleChatVisibility = () => {
-    setIsChatVisible(!isChatVisible);
-  };
 
   return (
     <ThemeProvider
@@ -28,8 +20,6 @@ function App() {
       <GlobalStyle />
       <ThemeToggle menuArr={["🌝", "🌚"]} />
       <RouterProvider router={routers} />
-      {isChatVisible && <Chat onClose={toggleChatVisibility} />}
-      <ChatButton onClick={toggleChatVisibility} />
     </ThemeProvider>
   );
 }
